trace-listen: export two functions for further use
Original do_listen() is the main process to let trace-cmd work as a
server, and receive client requests. For now, trace-cmd will only be a
server when we specify "trace-cmd listen", to collect tracing logs from
different hosts.

This patch exported do_listen() function to public, and renamed to
tracecmd_listen_start(). This would be a good framework when trace-cmd
is going to work as a server for other purpose. One parameter is added,
which is the service handler. When we want to extend supported services
for trace-cmd, we can provide a handler, and use the framework.

Also, as the first service handler to handle "trace-cmd listen",
process_client() is renamed to tracecmd_listen_process_client(), and
it's exported too, which is possibly be used in further patches.

Signed-off-by: Peter Xu <peterx@redhat.com>