trace-cmd-server: better handling for SIGINT
For trace-cmd-server, user will stop tracing on client side, we need to
capture the SIGINT and tell the server that we want to stop the remote
tracing. This procedure is reveresed comparing to trace-cmd-listen. We
need one more message (MSG_CLOSE) to notify the server side that we need
to stop current tracing.

Signed-off-by: Peter Xu <peterx@redhat.com>